+++ src/lib/tokenizer/merge_pair_core.cpp
@@ -9,13 +9,13 @@
 extern "C" {
 
 void merge_pair_core_parallel(
-    const int32_t* tokens_flat,
+    const uint16_t* tokens_flat,
     const int64_t* offsets,
     const size_t num_chunks,
-    const int32_t pair_L,
-    const int32_t pair_R,
-    const int32_t new_token,
-    int32_t* output_tokens_flat, // all -1 init. in-place change in this function
+    const uint16_t pair_L,
+    const uint16_t pair_R,
+    const uint16_t new_token,
+    uint16_t* output_tokens_flat, // all -1 init. in-place change in this function
     bool* output_filter, // all false init. in-place change in this function
     int64_t* output_tokens_lens // input tokens lens init. in-place change in this function
 ) {
